Refactor error types for SonarQube issues and Mulesoft standards
